CascadeProjects/windsurf-project/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic>=2.10.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
requests
torch
numpy
flask
gtts
pydub
speechrecognition
openai
openai-whisper
langchain-openai
langchain-community
pypdf
docarray
PyAudio
beautifulsoup4
python-magic
python-pptx
unstructured
python-docx
pandas
google-cloud-texttospeech
pygame
Flask-Mail
firebase_admin
flask_cors
langchain_chroma
langchain-yt-dlp
pytube
fastapi-mail
